Fix magnification type for Titan

+++ trident/slide_encoder_models/load.py
@@ -471,7 +471,7 @@ class TitanSlideEncoder(BaseSlideEncoder):
         return model, precision, embedding_dim
 
     def forward(self, batch, device='cuda'):
-        z = self.model.encode_slide_from_patch_features(batch['features'].to(device), batch['coords'].to(device), batch['attributes']['patch_size_level0'])        
+        z = self.model.encode_slide_from_patch_features(batch['features'].to(device), batch['coords'].to(device), batch['attributes']['patch_size_level0'].to(dtype=torch.int64))        
         return z
     
 class FeatherSlideEncoder(BaseSlideEncoder):
